# specs/001-ehr-frontend/contracts/api.yaml
openapi: 3.0.3
info:
  title: EHR Lite API
  description: Local-first oncology EHR backend API
  version: 1.0.0
  contact:
    name: EHR Lite Team
servers:
  - url: http://localhost:4000/api
    description: Local development server
...
paths:
...
  /patients/{id}/reports:
    get:
      summary: Get patient reports
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - name: type
          in: query
          schema:
            type: string
            enum: [pathology, imaging, lab, consultation, other]
          description: Filter by report type
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Report'
        '500':
          $ref: '#/components/responses/Error'

    post:
      summary: Upload new report
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/PatientId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - report_type
                - images
              properties:
                title:
                  type: string
                report_type:
                  type: string
                  enum: [pathology, imaging, lab, consultation, other]
                notes:
                  type: string
                report_date:
                  type: string
                  format: date
                diagnosis_id:
                  type: string
                  format: uuid
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Report uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Report'
        '400':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
  /reports/{id}:
    delete:
      summary: Delete report
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '404':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    PatientId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Patient UUID
...
    ReportId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Report UUID
  schemas:
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        diagnosis_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        report_type:
          type: string
          enum: [pathology, imaging, lab, consultation, other]
        notes:
          type: string
          nullable: true
        report_date:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        image_count:
          type: integer
          description: Number of attached images
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'